Add functionality to search Wikipedia
Add function (get_wiki_page) that takes a keyword as input and
downlaods the HTML content of that Wikipedia page.

Add child class (ParagraphParser) of HTMLParser that is able get the
contents within all <p> elements in the HTML.

Add function (find_keywords) that uses ParagraphParser to search
HTML for a pragraph containing the passed in keywords.